Removed charCodeChecker support
