+++ src/main/index.ts
@@ -15,4 +15,3 @@ export {Code, Var} from './code';
 export {createRule} from './rules/createRule';
 export {Rule, RuleHandler} from './rules/rule-types';
 export {Tokenizer} from './Tokenizer';
-export {InternalTaker} from './takers';
