+++ src/main/rules/index.ts
@@ -1,3 +1,4 @@
 export * from './compileRuleIterator';
 export * from './createRule';
 export * from './rule-types';
+export {toTakerFunction} from './rule-utils';
