+++ src/main/takers/internal-taker-types.ts
@@ -9,12 +9,10 @@ export interface InternalTaker extends TakerCodegen {
 
 const createType = Symbol;
 
-export const ALL_CHAR_CODE_CHECKER_TYPE = createType();
 export const ALL_CHAR_CODE_RANGE_TYPE = createType();
 export const ALL_CASE_SENSITIVE_TEXT_TYPE = createType();
 export const ALL_REGEX_TYPE = createType();
 export const ALL_GENERIC_TYPE = createType();
-export const CHAR_CODE_CHECKER_TYPE = createType();
 export const CHAR_CODE_RANGE_TYPE = createType();
 export const END_TYPE = createType();
 export const MAYBE_TYPE = createType();
@@ -26,16 +24,13 @@ export const CASE_SENSITIVE_TEXT_TYPE = createType();
 export const CASE_INSENSITIVE_TEXT_TYPE = createType();
 export const UNTIL_CASE_SENSITIVE_TEXT_TYPE = createType();
 export const UNTIL_CHAR_CODE_RANGE_TYPE = createType();
-export const UNTIL_CHAR_CODE_CHECKER_TYPE = createType();
 export const UNTIL_REGEX_TYPE = createType();
 export const UNTIL_GENERIC_TYPE = createType();
 
-export type ALL_CHAR_CODE_CHECKER_TYPE = typeof ALL_CHAR_CODE_CHECKER_TYPE;
 export type ALL_CHAR_CODE_RANGE_TYPE = typeof ALL_CHAR_CODE_RANGE_TYPE;
 export type ALL_CASE_SENSITIVE_TEXT_TYPE = typeof ALL_CASE_SENSITIVE_TEXT_TYPE;
 export type ALL_REGEX_TYPE = typeof ALL_REGEX_TYPE;
 export type ALL_GENERIC_TYPE = typeof ALL_GENERIC_TYPE;
-export type CHAR_CODE_CHECKER_TYPE = typeof CHAR_CODE_CHECKER_TYPE;
 export type CHAR_CODE_RANGE_TYPE = typeof CHAR_CODE_RANGE_TYPE;
 export type END_TYPE = typeof END_TYPE;
 export type MAYBE_TYPE = typeof MAYBE_TYPE;
@@ -47,6 +42,5 @@ export type CASE_SENSITIVE_TEXT_TYPE = typeof CASE_SENSITIVE_TEXT_TYPE;
 export type CASE_INSENSITIVE_TEXT_TYPE = typeof CASE_INSENSITIVE_TEXT_TYPE;
 export type UNTIL_CASE_SENSITIVE_TEXT_TYPE = typeof UNTIL_CASE_SENSITIVE_TEXT_TYPE;
 export type UNTIL_CHAR_CODE_RANGE_TYPE = typeof UNTIL_CHAR_CODE_RANGE_TYPE;
-export type UNTIL_CHAR_CODE_CHECKER_TYPE = typeof UNTIL_CHAR_CODE_CHECKER_TYPE;
 export type UNTIL_REGEX_TYPE = typeof UNTIL_REGEX_TYPE;
 export type UNTIL_GENERIC_TYPE = typeof UNTIL_GENERIC_TYPE;
